requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bardapi==0.1.33
beautifulsoup4==4.12.2
blinker==1.7.0
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
deep-translator==1.11.4
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.110.0
fonttools==4.39.4
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-resources==5.12.0
ipykernel==6.15.0
ipython==8.14.0
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==7.0.6
jupyter_core==5.3.0
keras==2.9.0
kiwisolver==1.4.4
langdetect==1.0.9
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.23.1
opencv-python==4.9.0.80
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==22.1.2
platformdirs==3.5.1
popen==0.1.20
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.24.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2022.1
pyzmq==19.0.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
setuptools==61.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.36.3
streamlit==1.32.2
tenacity==8.2.3
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.1
torch==1.12.0
tornado==6.1
traitlets==5.9.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.29.0
watchdog==4.0.0
wcwidth==0.2.6
wheel==0.37.1
zipp==3.15.0
